moving the channels_map to its own file
