+++ log_analyzer/src/static/js/application.js
@@ -1,5 +1,5 @@
 import { create_channels } from './channels.js';
-import { get_channel_color } from './utils.js';
+import { create_map } from './channels_map.js'
 
 $(document).ready(function(){
     let socket = io.connect('http://' + document.domain + ':' + location.port);
@@ -21,15 +21,7 @@ $(document).ready(function(){
         i++;
     });
     socket.on('channels_map', function(channels){
-        var nodes = create_nodes(channels);
-        var edges = create_edges(channels);
-        var container = document.getElementById("channels-map");
-        var data = {
-            nodes: nodes,
-            edges: edges
-        };
-        var options = {};
-        var network = new vis.Network(container, data, options);
+      create_map(channels);
    });
 });
 
@@ -115,23 +107,4 @@ function update_chart(chart, labels, data) {
     chart.data.datasets[0].data = data;
     chart.data.labels = labels;
     chart.update();
-}
-
-function create_nodes(arr) {
-    let ret = [];
-    arr.forEach(function(item, index, array) {
-        ret.push({id: index, value:12, shape: "box",
-                  color: get_channel_color(item), label: item});
-    })
-
-    return ret;
-}
-
-function create_edges(arr) {
-    let ret = [];
-    for (let i = 0; i < (arr.length - 1); i++) {
-        ret.push({from: i, to: i+1, arrows: "to"});
-    }
-
-    return ret;
 }
